refactor(drivers): update project version and remove unused functions
This commit removes several unused functions from the paw3903_opticflow.h file, such as `paw3903_spi_setup` and `paw3903_read_deltas`, indicating a cleanup of deprecated or redundant code. Additionally, it introduces new failure codes for the PAW3903 and VL53L1X sensors, expanding the error handling mechanism in the system.c file. A minor SPI configuration change is made in target.h, correcting the NSS pin assignment for improved hardware interaction. These changes enhance maintainability by reducing clutter and aligning the codebase with the current project state.

+++ src/main/drivers/system.h
@@ -8,7 +8,7 @@
  #  Created Date: Mon, 6th Oct 2025                                            #
  #  Brief:                                                                     #
  #  - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -  #
- #  Last Modified: Wed, 26th Nov 2025                                          #
+ #  Last Modified: Thu, 27th Nov 2025                                          #
  #  Modified By: AJ                                                            #
  #  - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -  #
  #  HISTORY:                                                                   #
@@ -30,7 +30,7 @@ uint32_t millis ( void );
 
 #ifdef TEST_ENABLE
 // failure
-void failureMode ( uint8_t mode );
+void failureMode ( uint16_t mode );
 #endif
 
 // bootloader/IAP
@@ -64,7 +64,8 @@ typedef enum {    // DD
   FAILURE_PAW3903,
   FAILURE_VL53L1X
 } failureMode_e;
-extern uint8_t failureFlag;
+
+extern uint16_t failureFlag;
 
 #ifdef __cplusplus
 }
